+++ tests/testthat/test-sf-from-raster.R
@@ -22,7 +22,7 @@ test_that("creation of polygons from raster works", {
 
 d <- raster::rasterize(lux, raster::raster(lux, res = 0.01))
 test_that("we can also qm_raster here", {
-  qm_rasterToPolygons(d) %>% 
+  polygonize(d) %>% 
     expect_named(c("ID_1", "NAME_1", "ID_2", "NAME_2", "AREA", "geometry")) %>% 
     expect_s3_class("sf")
 })