+++ NEWS.md
@@ -1,5 +1,7 @@
 # spex dev
 
+* new function 'buffer_extent' for whole-grain expansion
+
 * added S3 generic polygonize, alias of qm_rasterToPolygons
 
 * removed dependence on sf